pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pfq"
version = "0.1.0"
description = "A reasoning-focused personal task manager"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "click>=8.0",
    "pyperclip>=1.8",
    "pyyaml>=6.0",
    "textual>=0.60",
    "dateparser>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.scripts]
pfq = "pfq.__main__:cli"

[tool.setuptools.packages.find]
include = ["pfq*"]
